ex02の追加

+++ ex02/main.cpp
@@ -1,12 +1,14 @@
 #include <iostream>
-#include <Array.hpp>
+#include "Array.hpp"
+#include <cstdlib>
+#include <ctime>
 
 #define MAX_VAL 750
 int main(int, char**)
 {
     Array<int> numbers(MAX_VAL);
     int* mirror = new int[MAX_VAL];
-    srand(time(NULL));
+    srand(static_cast<unsigned int>(time(NULL)));
     for (int i = 0; i < MAX_VAL; i++)
     {
         const int value = rand();
